+++ stream.c
@@ -35,6 +35,10 @@
 #endif
 #endif
 
+#ifndef MPG_BUFFER_SIZE
+#define MPG_BUFFER_SIZE 128  // must be a power of 2
+#endif
+
 DCRAM static stream_rx_buffer_t rxbackup;
 
 typedef struct {
@@ -59,6 +63,17 @@ typedef struct stream_connection {
     struct stream_connection *next, *prev;
 } stream_connection_t;
 
+#if MPG_BUFFER_SIZE
+
+typedef struct {
+    volatile uint_fast16_t head;
+    uint8_t eol;
+    uint8_t data[MPG_BUFFER_SIZE];
+    char input[MPG_BUFFER_SIZE];
+} stream_mpg_buffer_t;
+
+#endif
+
 PROGMEM static const io_stream_properties_t null_stream = {
     .type = StreamType_Null,
     .instance = 0,
@@ -80,6 +95,9 @@ static struct {
     io_stream_t stream;
     stream_write_char_ptr write_char;
     stream_connection_flags_t flags;
+#if MPG_BUFFER_SIZE
+    stream_mpg_buffer_t *rx_buf;
+#endif
     on_rt_reports_added_ptr on_rt_reports_added;
     on_gcode_mode_changed_ptr on_gcode_mode_changed;
 } mpg;
@@ -620,6 +638,20 @@ static void stream_mpg_write (void *cmd)
 {
     switch((uintptr_t)cmd) {
 
+#if MPG_BUFFER_SIZE
+        case ASCII_LF:
+            {
+                status_code_t status;
+                if((status = *mpg.rx_buf->input == '$' ? system_execute_line(mpg.rx_buf->input, mpg.stream.write) : (*mpg.rx_buf->input ? Status_AccessDenied : Status_OK)) == Status_OK)
+                    mpg.stream.write("ok" ASCII_EOL);
+                else {
+                    mpg.stream.write("error:");
+                    mpg.stream.write(uitoa(status));
+                    mpg.stream.write(ASCII_EOL);
+                }
+            }
+            break;
+#endif
         case CMD_STATUS_REPORT_ALL:
             mpg.stream.report.flags.value = report_get_rt_flags_all().value;
             // no break
@@ -651,8 +683,28 @@ ISR_CODE bool ISR_FUNC(stream_mpg_check_enable)(uint8_t c)
             break;
 
         default:
-            protocol_enqueue_realtime_command(c);
-            if((c == CMD_CYCLE_START || c == CMD_CYCLE_START_LEGACY) && state_get() == STATE_IDLE)
+            if(!protocol_enqueue_realtime_command(c)) {
+#if MPG_BUFFER_SIZE
+                if(mpg.rx_buf) {
+
+                    uint16_t next_head = (mpg.rx_buf->head + 1) & (MPG_BUFFER_SIZE - 1);
+
+                    if(c == ASCII_LF || c == ASCII_CR) {
+                        mpg.rx_buf->eol = mpg.rx_buf->eol && c != mpg.rx_buf->eol ? '\0' : c;
+                        c = '\0';
+                    } else
+                        mpg.rx_buf->eol = '\0';
+
+                    if((mpg.rx_buf->data[mpg.rx_buf->head] = c))
+                        mpg.rx_buf->head = next_head;
+                    else if(mpg.rx_buf->eol) {
+                        mpg.rx_buf->head = 0;
+                        strcpy(mpg.rx_buf->input, (char *)mpg.rx_buf->data);
+                        task_add_immediate(stream_mpg_write, (void *)((uintptr_t)ASCII_LF));
+                    }
+                }
+#endif
+            } else if((c == CMD_CYCLE_START || c == CMD_CYCLE_START_LEGACY) && state_get() == STATE_IDLE)
                 report_add_realtime(Report_CycleStart);
             break;
     }
@@ -692,6 +744,9 @@ FLASHMEM bool stream_mpg_register (const io_stream_t *stream, bool rx_only, stre
         mpg.flags.is_mpg_tx = On;
         mpg.flags.mpg_control = Off;
 
+#if MPG_BUFFER_SIZE
+        mpg.rx_buf = calloc(sizeof(stream_mpg_buffer_t), 1);
+#endif
         if(write_char)
             mpg.stream.set_enqueue_rt_handler(write_char);
         else