+++ system.h
@@ -26,6 +26,7 @@
 #include "gcode.h"
 #include "probe.h"
 #include "alarms.h"
+#include "stream.h"
 #include "messages.h"
 #if NGC_EXPRESSIONS_ENABLE
 #include "vfs.h"
@@ -221,7 +222,8 @@ typedef union {
                  travel_changed          :1, //!< Set to true when maximum travel settings has changed.
                  is_homing               :1,
 				 is_parking			     :1, //!< Set to true when CMD_SAFETY_DOOR is received.
-                 unused                  :3;
+				 soft_estop :1,
+                 unused                  :2;
     };
 } system_flags_t;
 
@@ -295,7 +297,7 @@ typedef union {
         uint8_t noargs         :1, //!< Command does not handle arguments.
                 allow_blocking :1, //!< Command can be executed when blocking event is active.
                 help_fn        :1, //!< Command help is dynamic (output is via function call).
-                allow_mpg      :1, //!< Command can be executed from other streams than the current.
+                allow_redirect :1, //!< Command output can be to other other streams than the current.
                 unused         :4;
     };
 } sys_command_flags_t;
@@ -324,7 +326,7 @@ uint8_t system_n_axis (void);
 uint8_t system_axis_mask (void);
 uint8_t system_claim_axis (void);
 
-status_code_t system_execute_line (char *line);
+status_code_t system_execute_line (char *line, stream_write_ptr write);
 void system_execute_startup (void *data);
 void system_flag_wco_change (void);
 void system_convert_array_steps_to_mpos (float *position, int32_t *steps);